requirements.txt
streamlit>=1.45.0
requests>=2.32.0
Pillow>=10.0.0
bcrypt>=4.0.0
pandas>=2.0.0
plotly>=5.15.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
websockets>=12.0 